+++ src/canvas/SelectableCanvas.ts
@@ -293,7 +293,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
   protected declare _isCurrentlyDrawing: boolean;
   declare freeDrawingBrush?: BaseBrush;
   declare _activeObject?: FabricObject;
-  protected readonly _activeSelection: ActiveSelection;
+  protected _activeSelection: ActiveSelection;
 
   constructor(
     el?: string | HTMLCanvasElement,
@@ -598,7 +598,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
       control = !!corner && target.controls[corner],
       actionHandler =
         alreadySelected && control
-          ? control.getActionHandler(e, target, control)
+          ? control.getActionHandler(e, target, control)?.bind(control)
           : dragHandler,
       action = getActionFromCorner(alreadySelected, corner, e, target),
       origin = this._getOriginFromCorner(target, corner),
@@ -1086,6 +1086,12 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
     }
     this._activeObject = object;
 
+    if (object instanceof ActiveSelection && this._activeSelection !== object) {
+      this._activeSelection = object;
+      object.set('canvas', this);
+      object.setCoords();
+    }
+
     return true;
   }
 
@@ -1161,7 +1167,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
    */
   destroy() {
     // dispose of active selection
-    const activeSelection = this._activeSelection!;
+    const activeSelection = this._activeSelection;
     activeSelection.removeAll();
     // @ts-expect-error disposing
     this._activeSelection = undefined;